requirements.txt
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pylint==2.3.1
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
